+++ modules/payu/payment.php
@@ -1,6 +1,6 @@
 <?php
 /**
-*	ver. 1.9.1
+*	ver. 1.9.2
 *	PayU Payment Modules
 *	
 *	@copyright  Copyright 2012 by PayU