+++ modules/payu/validation.php
@@ -1,6 +1,6 @@
 <?php
 /**
- *	ver. 1.9
+ *	ver. 1.9.2
  *	PayU Payment Modules
  *
  *	@copyright  Copyright 2012 by PayU